Fixed some warnings in dump.c
Listen, it might seem excessive to fix all of those unused parameter warnings by voiding them, but it is the only thing that works, because they get called anonymously
(that is possible, right?), and so the function that is calling them just passes all of that information because other ones that might need it also get called the same
way. Just so you know. I tried my best to not do that.