Achievements!
This is the first of my commits to adding achievements to BingeHack4. So
far, I have rewritten the achieve.c to not require a MySQL database, but
be much more portable. I still need to make it so it saves between games
though. Just realized that one. Maybe I can set it up so that it gets
all of the data on achievements from text files or something. Also, I
need to find out how I can actually save your specific progress between
games.

Other than that, I've been working on putting the hooks into the code
for BingeHack4, mostly it is the same as BingeHack, but some of them
need to be re-thought a little. This is my first commit for achievements
of many, because I just realized that I should actually commit my work
as I go. Yay!