fixing the totally stupid issues I made
this commit should actually work, like, I mean, compile at all. The
issue being that I need to figure out why it isn't linking. After some
thinking, my suspision is that achieve isn't in the Makefile, but I am
going to test that out by commiting this and moving it over to rancor,
where the Makefile is presumably correct, and there I am going to try to
make it. If it works, I am going to diff the files and fix mine. If it
doesn't, I am going to bang my head against a wall.